backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
chromadb==0.4.22
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pypdf==3.17.1
pandas==2.1.0
openpyxl==3.1.2
pyarrow==14.0.1
aiofiles==23.2.1
pytest==7.4.0
httpx==0.25.0
itsdangerous==2.1.2
